Auth with JWT clean version
