--- src/main/java/com/develhope/spring/user/configurations/ApplicationConfiguration.java
@@ -1,4 +1,0 @@
-package com.develhope.spring.user.configurations;
-
-public class ApplicationConfiguration {
-}
