--- src/main/java/com/develhope/spring/user/controller/AuthController.java
@@ -1,57 +1,0 @@
-package com.develhope.spring.user.controller;
-
-import com.develhope.spring.user.dto.LoginDto;
-import com.develhope.spring.user.entity.Account;
-import com.develhope.spring.user.entity.LoginResponse;
-import com.develhope.spring.user.service.JwtService;
-import com.develhope.spring.user.service.UserService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/auth")
-public class AuthController {
-//    @Autowired
-//    private AuthenticationManager authenticationManager;
-
-    private final UserService userService;
-
-    private final JwtService jwtService;
-
-    public AuthController(UserService userService, JwtService jwtService) {
-        this.userService = userService;
-        this.jwtService = jwtService;
-    }
-
-    @PostMapping("/login")
-    public ResponseEntity<LoginResponse> authenticate(@RequestBody LoginDto loginUserDto) {
-        Account authenticatedUser = userService.authenticate(loginUserDto);
-
-        String jwtToken = jwtService.generateToken(authenticatedUser);
-
-        LoginResponse loginResponse = new LoginResponse();
-        loginResponse.setToken(jwtToken);
-        loginResponse.setExpiresIn(jwtService.getExpirationTime());
-
-        return ResponseEntity.ok(loginResponse);
-    }
-
-    //    @PostMapping("/signin")
-//    public ResponseEntity<String> authenticateUser(@RequestBody LoginDto loginDto){
-//        Authentication authentication = authenticationManager.authenticate(new UsernamePasswordAuthenticationToken(
-//                loginDto.getUsernameOrEmail(), loginDto.getPassword()));
-//
-//        SecurityContextHolder.getContext().setAuthentication(authentication);
-//        return new ResponseEntity<>("User signed-in successfully!.", HttpStatus.OK);
-//    }
-
-}
