+++ src/main/java/com/develhope/spring/user/controller/UserController.java
@@ -1,7 +1,10 @@
 package com.develhope.spring.user.controller;
 
+import com.develhope.spring.user.dto.LoginDto;
 import com.develhope.spring.user.dto.RegistrationDto;
 import com.develhope.spring.user.entity.Account;
+import com.develhope.spring.user.entity.LoginResponse;
+import com.develhope.spring.user.service.JwtService;
 import com.develhope.spring.user.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -12,8 +15,14 @@ import java.util.List;
 @RestController
 @RequestMapping("/user")
 public class UserController {
-    @Autowired
-    private UserService userService;
+
+    private final UserService userService;
+    private final JwtService jwtService;
+
+    public UserController(UserService userService, JwtService jwtService) {
+        this.userService = userService;
+        this.jwtService = jwtService;
+    }
 
     @GetMapping
     public List<Account> getAllUsers() {
@@ -24,6 +33,19 @@ public class UserController {
     public ResponseEntity<Account> createProfile(@RequestBody RegistrationDto registrationDto) {
         Account createdAccount = userService.create(registrationDto);
         return ResponseEntity.ok(createdAccount);
+    }
+
+    @PostMapping("/login")
+    public ResponseEntity<LoginResponse> authenticate(@RequestBody LoginDto loginUserDto) {
+        Account authenticatedUser = userService.authenticate(loginUserDto);
+
+        String jwtToken = jwtService.generateToken(authenticatedUser);
+
+        LoginResponse loginResponse = new LoginResponse();
+        loginResponse.setToken(jwtToken);
+        loginResponse.setExpiresIn(jwtService.getExpirationTime());
+
+        return ResponseEntity.ok(loginResponse);
     }
 
     @DeleteMapping("/delete/{id}")
