+++ src/main/java/com/develhope/spring/user/entity/LoginResponse.java
@@ -1,17 +1,12 @@
 package com.develhope.spring.user.entity;
 
-import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Data;
-import lombok.NoArgsConstructor;
 
-@NoArgsConstructor
 @Data
 public class LoginResponse {
     private String token;
 
     private long expiresIn;
-
 
 
 }
