+++ src/main/java/com/develhope/spring/user/entity/UserKind.java
@@ -3,9 +3,7 @@ package com.develhope.spring.user.entity;
 import org.springframework.security.core.GrantedAuthority;
 
 public enum UserKind implements GrantedAuthority {
-    BUYER,
-    SELLER,
-    ADMIN;
+    BUYER, SELLER, ADMIN;
 
     @Override
     public String getAuthority() {
