+++ src/main/java/com/develhope/spring/user/service/CustomUserDetailService.java
@@ -1,7 +1,5 @@
 package com.develhope.spring.user.service;
 
-
-import com.develhope.spring.user.entity.Account;
 import com.develhope.spring.user.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -22,3 +20,4 @@ public class CustomUserDetailService implements UserDetailsService {
         );
     }
 }
+
