+++ src/main/java/com/develhope/spring/vehicles/controller/VehicleController.java
@@ -6,6 +6,8 @@ import com.develhope.spring.vehicles.service.VehicleService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -51,6 +53,7 @@ public class VehicleController {
 
     @GetMapping
     public ResponseEntity<List<Vehicle>> findAll() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         return ResponseEntity
                 .status(HttpStatus.FOUND)
                 .body(vehicleService
